+++ omop_alchemy/cdm/model/vocabulary/concept.py
@@ -27,8 +27,8 @@ class Concept(
     concept_class_id: so.Mapped[str] = so.mapped_column(sa.ForeignKey("concept_class.concept_class_id"), nullable=False, index=True)
     standard_concept: so.Mapped[Optional[str]] = so.mapped_column(sa.String(1))
     concept_code: so.Mapped[str] = so.mapped_column(sa.String(50), nullable=False)
-    valid_start_date: so.Mapped[date] = so.mapped_column(nullable=False)
-    valid_end_date: so.Mapped[date] = so.mapped_column(nullable=False)
+    valid_start_date: so.Mapped[date] = so.mapped_column(sa.Date(), nullable=False)
+    valid_end_date: so.Mapped[date] = so.mapped_column(sa.Date(), nullable=False)
     invalid_reason: so.Mapped[Optional[str]] = so.mapped_column(sa.String(1))
 
 class ConceptContext(ReferenceContext):